test open graph

+++ api/src/models/util/functionDB/CodeDiscountDb.js
@@ -1,14 +1,13 @@
 require('../../../DB');
 const { getUser } = require('../../../routes/services/users.services');
 const CodeDiscountModel = require('../../DB/CodeDiscount');
+const formatDate = require('../helpers/formatDate');
 
 const createCodeDiscount = async (data) => {
   try {
-    if (data.idCreator) {
-      const user = await getUser(data.idCreator);
-      user.availableCredit = user.availableCredit - data.value;
-      await user.save();
-    }
+    const user = await getUser(data.idCreator);
+    user.availableCredit = user.availableCredit - data.value;
+    await user.save();
     const newCodeDiscount = new CodeDiscountModel(data);
     return await newCodeDiscount.save();
   } catch (error) {
@@ -47,8 +46,10 @@ const getListCodeDiscountByCreator = async (id) => {
 const updateCodeDiscount = async (id, value, percentage, quotas) => {
   try {
     const codeDiscount = await CodeDiscountModel.findById(id);
-
     const userCodeDiscount = await getUser(codeDiscount.idCreator);
+
+    if (value > userCodeDiscount.availableCredit) throw new Error('No tienes suficiente saldo');
+    if (value === 0 || value < 0) throw new Error('Valor invalido');
 
     let auxValue = 0;
     if (value > codeDiscount.value) {
@@ -72,10 +73,34 @@ const updateCodeDiscount = async (id, value, percentage, quotas) => {
 
 const deleteCodeDiscountById = async (id) => {
   try {
+    const codeDiscount = await CodeDiscountModel.findById(id);
+
+    console.log({
+      codeDiscount,
+    });
+
+    if (codeDiscount.isRedimeed) throw new Error('Este código ya ha sido utilizado.');
+
     const codeDiscountDeleted = await CodeDiscountModel.findByIdAndDelete(id);
     const userCreateCodeDiscount = await getUser(codeDiscountDeleted.idCreator);
     userCreateCodeDiscount.availableCredit = userCreateCodeDiscount.availableCredit + codeDiscountDeleted.value;
     await userCreateCodeDiscount.save();
+  } catch (error) {
+    throw new Error(error.message);
+  }
+};
+
+const redimeedCodeDiscount = async (idCode, idUser) => {
+  try {
+    const codeDiscount = await CodeDiscountModel.findById(idCode);
+
+    if (codeDiscount.isRedimeed) throw new Error('Este codigo ya ha sido redimido.');
+
+    const user = await getUser(idUser);
+    codeDiscount.dateRedimeed = formatDate(new Date());
+    codeDiscount.userRedimeed = user.nickname;
+    codeDiscount.isRedimeed = true;
+    await codeDiscount.save();
   } catch (error) {
     throw new Error(error.message);
   }
@@ -88,4 +113,5 @@ module.exports = {
   updateCodeDiscount,
   deleteCodeDiscountById,
   getListCodeDiscountByCreator,
+  redimeedCodeDiscount,
 };
